+++ packages/rate-limited-axios/lib/RateLimitedAxios.ts
@@ -1,23 +1,13 @@
 import originalAxios, {
-  AxiosError,
   AxiosRequestConfig,
-  AxiosResponse,
   InternalAxiosRequestConfig,
 } from 'axios';
+import { Semaphore } from 'await-semaphore';
+import { RateLimiterMemory } from 'rate-limiter-flexible';
 import { RateLimitedAxiosConfig } from './config';
-import { Rule } from './types';
-
-declare module 'axios' {
-  export interface InternalAxiosRequestConfig {
-    meta: { release: (() => void) | undefined; startedTime: number };
-  }
-}
 
 class RateLimitedAxios extends originalAxios.Axios {
-  protected static timeoutIndicesPerRelease = new Map<
-    () => void,
-    ReturnType<typeof setTimeout>
-  >();
+  protected static consumedData: { [key: string]: number } = {};
 
   constructor(config?: AxiosRequestConfig) {
     super(
@@ -26,93 +16,42 @@ class RateLimitedAxios extends originalAxios.Axios {
         config || {}
       )
     );
-    this.interceptors.request.use(RateLimitedAxios.interceptorForRequest);
-    this.interceptors.response.use(
-      RateLimitedAxios.interceptorForResponse,
-      RateLimitedAxios.interceptorForResponseError
-    );
+    this.interceptors.request.use(RateLimitedAxios.interceptor);
   }
-
-  /**
-   * Releases the request queue associated with the given Axios request configuration.
-   * @param config
-   * @returns
-   */
-  protected static releaseQueue = (config: InternalAxiosRequestConfig) => {
-    const url = config.url ?? '';
-    const rule = RateLimitedAxios.getUrlRule(url);
-
-    if (rule && config.meta.release) {
-      const key = rule.pattern.toString();
-      const release = config.meta['release'];
-      let releaseTime =
-        rule.throttleWindow * 1000 - (Date.now() - config.meta.startedTime);
-      if (releaseTime < 0) releaseTime = 0;
-      setTimeout(() => {
-        clearTimeout(RateLimitedAxios.timeoutIndicesPerRelease.get(release));
-        RateLimitedAxios.timeoutIndicesPerRelease.delete(release);
-        // release the locked queue
-        release();
-      }, releaseTime);
-    }
-  };
 
   /**
    * This function manages rate limiting for requests by matching URLs against regex patterns.
    * @param config
    * @returns
    */
-  protected static interceptorForRequest = async (
-    config: InternalAxiosRequestConfig
-  ) => {
-    const url = config.url ?? '';
-    const rule = RateLimitedAxios.getUrlRule(url);
+  protected static interceptor = async (config: InternalAxiosRequestConfig) => {
+    const url = config.baseURL
+      ? originalAxios.getUri({ baseURL: config.baseURL, url: config.url })
+      : config.url ?? '';
+    const [limiter, pattern, semaphore] =
+      RateLimitedAxios.getLimiterAndPatternOfUrl(url);
 
-    if (!rule) return config;
+    if (!limiter) return config;
 
-    const key = rule.pattern.toString();
-    const release = await rule.semaphore.acquire();
-    config.meta = { release: release, startedTime: Date.now() };
+    const key = pattern.toString();
+    const release = await semaphore.acquire();
 
     try {
-      RateLimitedAxios.timeoutIndicesPerRelease.set(
-        release,
-        setTimeout(() => {
-          RateLimitedAxiosConfig.getLogger().debug(
-            `The response time has exceeded the defined limit for the ${key} URL pattern`
-          );
-          RateLimitedAxios.releaseQueue(config);
-        }, rule.timeout * 1000)
-      );
-    } catch (err) {
-      RateLimitedAxiosConfig.getLogger().error(
-        `Error on the RateLimitedAxios.interceptorForRequest occurred: ${err}`
-      );
+      if ((await limiter.get(key))?.remainingPoints === 0) {
+        RateLimitedAxiosConfig.getLogger().info(
+          `Rate limit exceeded for "${pattern}" url pattern, waiting for ${
+            (await limiter.get(key))!.msBeforeNext
+          }ms`
+        );
+        const msBeforeNext = (await limiter.get(key))!.msBeforeNext;
+        await new Promise((f) => setTimeout(f, msBeforeNext));
+      }
+      await limiter.consume(key);
+    } finally {
       release();
-      throw err;
     }
 
     return config;
-  };
-
-  /**
-   * Axios response interceptor that triggers queue release logic.
-   * @param response
-   * @returns
-   */
-  protected static interceptorForResponse = (response: AxiosResponse) => {
-    this.releaseQueue(response.config);
-    return response;
-  };
-
-  /**
-   * Axios error interceptor that handles failed responses by releasing the request queue.
-   * @param error
-   * @returns
-   */
-  protected static interceptorForResponseError = (error: AxiosError) => {
-    if (error.config) this.releaseQueue(error.config);
-    return Promise.reject(error);
   };
 
   /**
@@ -120,10 +59,17 @@ class RateLimitedAxios extends originalAxios.Axios {
    * @param url
    * @returns
    */
-  protected static getUrlRule = (url: string): Rule | undefined => {
-    for (const rule of RateLimitedAxiosConfig.getRules()) {
-      if (rule.pattern.test(url)) return rule;
+  protected static getLimiterAndPatternOfUrl = (
+    url: string
+  ): [RateLimiterMemory, RegExp, Semaphore] | [null, null, null] => {
+    for (const {
+      pattern,
+      limiter,
+      semaphore,
+    } of RateLimitedAxiosConfig.getRules()) {
+      if (pattern.test(url)) return [limiter, pattern, semaphore];
     }
+    return [null, null, null];
   };
 
   /**
@@ -152,4 +98,4 @@ const create = (config: AxiosRequestConfig = {}) => {
   return axiosInstance;
 };
 
-export { create, RateLimitedAxios, RateLimitedAxiosConfig };
+export { RateLimitedAxios, RateLimitedAxiosConfig, create };