+++ packages/rate-limited-axios/lib/config.ts
@@ -1,3 +1,4 @@
+import { RateLimiterMemory } from 'rate-limiter-flexible';
 import { AbstractLogger, DummyLogger } from '@rosen-bridge/abstract-logger';
 import { Rule } from './types';
 import { Semaphore } from 'await-semaphore';
@@ -15,15 +16,16 @@ export class RateLimitedAxiosConfig {
   public static addRule(
     pattern: string,
     rateLimit: number,
-    throttleWindow: number,
-    timeout: number
+    throttleWindow: number
   ) {
     RateLimitedAxiosConfig.removeRule(pattern);
     RateLimitedAxiosConfig.limitRules.push({
       pattern: new RegExp(pattern),
+      limiter: new RateLimiterMemory({
+        points: rateLimit,
+        duration: throttleWindow,
+      }),
       semaphore: new Semaphore(rateLimit),
-      throttleWindow: throttleWindow,
-      timeout: timeout,
     });
   }